+++ attest/key_tspi.go
@@ -12,7 +12,7 @@
 // License for the specific language governing permissions and limitations under
 // the License.
 
-// +build linux,!gofuzz
+// +build linux,tspi,!gofuzz
 
 package attest
